reorganize and comment
